Ideas for more commands/instructions
I added some additional commands that I think might be interesting.  Let me know if
any of them seem interesting asis or if you'd like to tweak how they work.

Here's the new ones included in the code in the PR.  The B/b/E/e ones are to make
it possible to copy things from one stack to another without having to use the
backup cell.  And the Q/q ones let you treat the stack like a wheel you can spin
in other direction.

_ : Integer division

A : bitwise "and" top two stack entries

O : bitwise "or" top two stack entries

X : bitwise "xor" top two stack entries

B : push a copy of the top of the stack onto the stack to the left

b : push a copy of the bottom of the stack onto the stack to the left

E : push a copy of the top of the stack onto the stack to the right

e : push a copy of the bottom of the stack onto the stack to the right

Y : "yank" an entry from the stack, push it on the top (like "I" but it deletes what it copies)

q : pops a number off the top of the stack, rotates the stack down that many times

Q : pops a number off the top of the stack, rotates the stack up that many times