Fix brackets in comments and character/string literals
This commit fixes three related problems with brackets in comments and strings:

First problem is that regex code which removes strings and comments for the match_brackets function
doesn't handle all the situations correctly. For example, this program:

 #"
 ["

passes match_brackets test.

Second problem is that { immediately following #'" is not handled correctly:

 #{0u

this program prints 0.

Third problem is that comments and strings are not taken into account while looking for matching bracket,
this program jumps inside a string and prints 93 (should print nothing):

["]"]

This last problem is the hardest.

I tried to handle these issues directly, but it was a mess due to the number of different possibilities
in which strings, characters, comments and brackets interact with each other. The code asks for proper
tokenizer, but it is too much for such a simple language.

So I propose this solution: add a preprocess step, which removes comments and replaces brackets inside
string and character literals with another chars. I use U+FDD0..U+FDD5 for this purpose, they are
designated by Unicode for internal application usage.

After this we don't have to worry about brackets at all - they just work.
All we have to do is change U+FDD0..U+FDD5 back to ()[]{} when reading string/char literals.

It is hard to create proper regex that removes comments while taking string/character literals into account.
On the other hand, the preprocess step already has to deal with all these, so it is more natural and easy
to remove comments manually. So I removed the regexes, everything is handled by the preprocess function now.